+++ src/components/WidgetForm/SimpleSearch/SimpleSearchForm.js
@@ -4,10 +4,12 @@ import PropTypes from 'prop-types';
 import { useForm } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
 
+import {
+  AccordionSet,
+} from '@folio/stripes/components';
+
 import SimpleSearchFilterArray from './filters/SimpleSearchFilterArray';
 import SimpleSearchResults from './results/SimpleSearchResults';
-
-import SimpleSearchSort from './sort/SimpleSearchSort';
 import SimpleSearchConfigurableProperties from './configurableProperties/SimpleSearchConfigurableProperties';
 
 const SimpleSearchForm = ({
@@ -43,34 +45,40 @@ const SimpleSearchForm = ({
     }
   }, [change, defChanged, toggleDefChange]);
 
+  const initialAccordionState = {
+    filters: true,
+    results: true,
+    sort: true
+  };
+
   return (
     <>
-      <SimpleSearchConfigurableProperties
-        configurableProperties={configurableProperties}
-      />
-      <FieldArray
-        addButtonId="simple-search-form-add-filter-button"
-        addLabelId="ui-dashboard.simpleSearchForm.filters.addFilter"
-        component={SimpleSearchFilterArray}
-        data={{
-          filterColumns
-        }}
-        deleteButtonTooltipId="ui-dashboard.simpleSearchForm.filters.removeFilter"
-        headerId="ui-dashboard.simpleSearchForm.filters"
-        id="simple-search-form-filters"
-        name="filterColumns"
-      />
-      <SimpleSearchResults
-        data={{
-          resultColumns
-        }}
-        id="simple-search-form-results"
-      />
-      <SimpleSearchSort
-        data={{
-          sortColumns
-        }}
-      />
+      <AccordionSet initialStatus={initialAccordionState}>
+        {/* This component now only displays url link stuff, consider renaming */}
+        <SimpleSearchConfigurableProperties
+          configurableProperties={configurableProperties}
+        />
+        <FieldArray
+          addButtonId="simple-search-form-add-filter-button"
+          addLabelId="ui-dashboard.simpleSearchForm.filters.addFilter"
+          component={SimpleSearchFilterArray}
+          data={{
+            filterColumns
+          }}
+          deleteButtonTooltipId="ui-dashboard.simpleSearchForm.filters.removeFilter"
+          headerId="ui-dashboard.simpleSearchForm.filters"
+          id="simple-search-form-filters"
+          name="filterColumns"
+        />
+        <SimpleSearchResults
+          data={{
+            resultColumns,
+            configurableProperties,
+            sortColumns
+          }}
+          id="simple-search-form-results"
+        />
+      </AccordionSet>
     </>
   );
 };