+++ src/components/WidgetForm/SimpleSearch/configurableProperties/SimpleSearchConfigurableProperties.js
@@ -10,32 +10,11 @@ import {
 
 const SimpleSearchConfigurableProperties = ({
   configurableProperties: {
-    numberOfRows = {},
     urlLink = {}
   } = {}
 }) => {
   return (
     <>
-      <Field
-        defaultValue={numberOfRows.defValue}
-        name="configurableProperties.numberOfRows"
-      >
-        {({ input }) => {
-          if (numberOfRows.configurable) {
-            return (
-              <TextField
-                {...input}
-                data-testid="simple-search-configurable-properties-number-of-rows"
-                id="simple-search-configurable-properties-number-of-rows"
-                label={<FormattedMessage id="ui-dashboard.simpleSearchForm.configurableProperties.numberOfRows" />}
-              />
-            );
-          }
-          // We know that if numberOfRows is non-configurable then it MUST have a defValue
-          // We still want that to be submitted, but no field to render on the form
-          return null;
-        }}
-      </Field>
       { (urlLink.configurable || urlLink.defValue) &&
         // If urlLink is non configurable and has no defValue then we don't need it on the form
         <Field