+++ src/components/WidgetForm/SimpleSearch/sort/SimpleSearchSort.js
@@ -5,13 +5,11 @@ import { FormattedMessage } from 'react-intl';
 import { Field, useForm } from 'react-final-form';
 import {
   Col,
-  Headline,
   KeyValue,
-  Row,
   Select
 } from '@folio/stripes/components';
 
-const SimpleSearchSort = ({ data: { sortColumns = [] } = {} }) => {
+const SimpleSearchSort = ({ data: { sortColumns } = {} }) => {
   const { change } = useForm();
 
   // Do this at the top of the form unconditionally for hook reasons
@@ -53,53 +51,44 @@ const SimpleSearchSort = ({ data: { sortColumns = [] } = {} }) => {
   const selectifiedSortDirs = selectedSortCol?.sortTypes?.map(st => ({ value: st, label: st })) || [];
 
   return (
-    <KeyValue
-      data-testid="simple-search-sort"
-      label={
-        <Headline margin="x-small" size="medium" tag="h2">
-          <FormattedMessage id="ui-dashboard.simpleSearchForm.sort" />
-        </Headline>
-      }
-    >
-      <Row>
-        <Col xs={6}>
-          <KeyValue
-            data-testid="simple-search-sort-sort-by"
-            label={<FormattedMessage id="ui-dashboard.simpleSearchForm.sort.sortBy" />}
-          >
-            <Field
-              component={Select}
-              dataOptions={selectifiedSortColumns}
-              defaultValue={sortColumns[0].name}
-              name="sortColumn.name"
-              onChange={e => {
-                setSSC(sortColumns.find(sc => sc.name === e.target.value));
-                /*
-                 * This means you can't set sort order THEN sortBy,
-                 * but this is because in the definition sort direction belongs to the sortBy
-                */
-                change('sortColumn.sortType', selectedSortCol.sortTypes[0]);
-                change('sortColumn.name', e.target.value);
-              }}
-            />
-          </KeyValue>
-        </Col>
-        <Col xs={6}>
-          <KeyValue
-            data-testid="simple-search-sort-sort-direction"
-            label={<FormattedMessage id="ui-dashboard.simpleSearchForm.sort.sortDirection" />}
-          >
-            <Field
-              component={Select}
-              dataOptions={selectifiedSortDirs}
-              defaultValue={sortColumns[0].sortTypes[0]}
-              disabled={selectifiedSortDirs.length <= 1}
-              name="sortColumn.sortType"
-            />
-          </KeyValue>
-        </Col>
-      </Row>
-    </KeyValue>
+    <>
+      <Col xs={3}>
+        <KeyValue
+          data-testid="simple-search-sort-sort-by"
+          label={<FormattedMessage id="ui-dashboard.simpleSearchForm.sort.sortBy" />}
+        >
+          <Field
+            component={Select}
+            dataOptions={selectifiedSortColumns}
+            defaultValue={sortColumns[0].name}
+            name="sortColumn.name"
+            onChange={e => {
+              setSSC(sortColumns.find(sc => sc.name === e.target.value));
+              /*
+                * This means you can't set sort order THEN sortBy,
+                * but this is because in the definition sort direction belongs to the sortBy
+              */
+              change('sortColumn.sortType', selectedSortCol.sortTypes[0]);
+              change('sortColumn.name', e.target.value);
+            }}
+          />
+        </KeyValue>
+      </Col>
+      <Col xs={3}>
+        <KeyValue
+          data-testid="simple-search-sort-sort-direction"
+          label={<FormattedMessage id="ui-dashboard.simpleSearchForm.sort.sortDirection" />}
+        >
+          <Field
+            component={Select}
+            dataOptions={selectifiedSortDirs}
+            defaultValue={sortColumns[0].sortTypes[0]}
+            disabled={selectifiedSortDirs.length <= 1}
+            name="sortColumn.sortType"
+          />
+        </KeyValue>
+      </Col>
+    </>
   );
 };