+++ src/index.js
@@ -4,6 +4,7 @@ import { Route } from '@folio/stripes/core';
 import PropTypes from 'prop-types';
 
 const Settings = lazy(() => import('./settings'));
+const DashboardsRoute = lazy(() => import('./routes/DashboardsRoute'));
 const DashboardRoute = lazy(() => import('./routes/DashboardRoute'));
 const DashboardOrderRoute = lazy(() => import('./routes/DashboardOrderRoute'));
 const WidgetCreateRoute = lazy(() => import('./routes/WidgetCreateRoute'));
@@ -32,7 +33,7 @@ class App extends React.Component {
           <Route component={WidgetCreateRoute} path={`${path}/:dashName/create`} />
           <Route component={DashboardOrderRoute} path={`${path}/:dashName/editOrder`} />
           <Route component={DashboardRoute} path={`${path}/:dashName`} />
-          <Route component={DashboardRoute} path={path} />
+          <Route component={DashboardsRoute} path={path} />
         </Switch>
       </Suspense>
     );